Moved elementary stand alone functions
