+++ src/elementary.hpp
@@ -11,8 +11,8 @@
 #include "elementary/ReactionID.hpp"
 
 // free functions to produce new identifiers
-#include "elementary/src/absorb.hpp"
-#include "elementary/src/emit.hpp"
-#include "elementary/src/resolve.hpp"
-#include "elementary/src/fromEndfReactionNumber.hpp"
-#include "elementary/src/toEndfReactionNumber.hpp"
+#include "elementary/absorb.hpp"
+#include "elementary/emit.hpp"
+#include "elementary/resolve.hpp"
+#include "elementary/fromEndfReactionNumber.hpp"
+#include "elementary/toEndfReactionNumber.hpp"
