+++ src/elementary/absorb.hpp
