+++ src/elementary/emit.hpp
